% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iCF_SUBGROUP_ANALYSIS.R
\name{ANA_SUBGROUP}
\alias{ANA_SUBGROUP}
\title{Function that calcualts observed treatment effect size}
\usage{
ANA_SUBGROUP(subgroup)
}
\arguments{
\item{subgroup}{the database (subgroup) in a list of stratified population accroding to subgroup decision}
}
\value{
The ATE (IPTW), ATT (SMR), and crude treatment effect (i.e., Delta Y)
}
\description{
Function that calcualts observed treatment effect size
}
