% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iCF_SUBGROUP_ANALYSIS.R
\name{CATE_MAX_SG}
\alias{CATE_MAX_SG}
\title{Function that identify subgroup with the largest CATE defined by:
\enumerate{
\item binary outcome: Q(A)=(P(Y=1|T=1, X∈A) - P(Y=1|T=0, X∈A)) – (P(Y=1|T=1) - P(Y=1|T=0))
\item continuous:     Q(A)=(E(Y=1|T=1, X∈A) - E(Y=1|T=0, X∈A)) – (E(Y=1|T=1) - E(Y=1|T=0))
}}
\usage{
CATE_MAX_SG(DltY.data)
}
\arguments{
\item{DltY.data}{calculated dltY}
}
\value{
the subgroup leading to the largest Q(A)
}
\description{
Function that identify subgroup with the largest CATE defined by:
\enumerate{
\item binary outcome: Q(A)=(P(Y=1|T=1, X∈A) - P(Y=1|T=0, X∈A)) – (P(Y=1|T=1) - P(Y=1|T=0))
\item continuous:     Q(A)=(E(Y=1|T=1, X∈A) - E(Y=1|T=0, X∈A)) – (E(Y=1|T=1) - E(Y=1|T=0))
}
}
