% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iCF_SUBGROUP_ANALYSIS.R
\name{CATE_SG}
\alias{CATE_SG}
\title{Function that calculate per-subgroup treatment effect:}
\usage{
CATE_SG(dat, SGdecision)
}
\arguments{
\item{dat}{data set for overall population}

\item{SGdecision}{subgroup decision}
}
\value{
the subgroup leading to the largest Q(A)
}
\description{
Function that calculate per-subgroup treatment effect:
}
