% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iCF_SG_PIPELINE.R
\name{CF}
\alias{CF}
\title{This function is the wrapper function to run raw causal forest with predefined min.node.size = round(nrow(Train)/depth, 0),}
\usage{
CF(depth, treeNo, tunepara)
}
\arguments{
\item{depth}{parameter to tune min.node.size = round(nrow(Train)/depth, 0),}

\item{treeNo}{number of trees in each forest}

\item{tunepara}{request wihout tuning}
}
\value{
the key results from raw causal forest
}
\description{
This function is the wrapper function to run raw causal forest with predefined min.node.size = round(nrow(Train)/depth, 0),
}
