% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iCF_SUBGROUP_DECISION.R
\name{CF_GROUP_DECISION}
\alias{CF_GROUP_DECISION}
\title{#Function that obtain MODELS from subgroup decisions at different depths from CF}
\usage{
CF_GROUP_DECISION(HTE_P_cf.raw, dat, method_CF, outcome_type_at, P_threshold)
}
\arguments{
\item{dat}{the dataset for group lasso}

\item{method_CF}{actall outcome or transformed outcome, iCFCV funciton will proivde results for both}

\item{HE_P_value}{raw heterogeneity test P-value}

\item{formula_gl}{formula for group lasso}

\item{method_deci}{method for selecting "cv" or "shrink"}
}
\value{
the data for group lasso
}
\description{
#Function that obtain MODELS from subgroup decisions at different depths from CF
}
