% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iCF_TREE_build.R
\name{CF_RAW_key}
\alias{CF_RAW_key}
\title{This function is the wrapper function to run raw causal forest}
\usage{
CF_RAW_key(Train_cf, min.split.var, variable_type, hdpct)
}
\arguments{
\item{Train_cf}{training data}

\item{min.split.var}{minimum number of variables to split population}
}
\value{
the key results from raw causal forest
}
\description{
This function is the wrapper function to run raw causal forest
}
