% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iCF_grplasso.R
\name{cv.grplasso}
\alias{cv.grplasso}
\title{Function that run cross-validation for group lasso (grplasso) using a formula (formula_grplasso) designed to CF,
BE CAREFUL!!! Need to revise for classification.}
\usage{
cv.grplasso(dat, lambda_grplasso, formula_grplasso, contrast)
}
\arguments{
\item{dat}{the dataset}

\item{lambda_grplasso}{a grid of penalty parameter from 0 to lambda_max}
}
\value{
the list of cross-validation results and the lambda leading to minimor.
}
\description{
Function that run cross-validation for group lasso (grplasso) using a formula (formula_grplasso) designed to CF,
BE CAREFUL!!! Need to revise for classification.
}
