% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iCF_MAJORITY_VOTE.R
\name{CV_SG_MAJORITY}
\alias{CV_SG_MAJORITY}
\title{This function finds the majority voted subgroup decision across CV}
\usage{
CV_SG_MAJORITY(vote_D_subgroup.L, stability_D_T_r, K)
}
\arguments{
\item{vote_D_subgroup.L}{list of vote_D_subgroup across CV}
}
\value{
The majority voted subgroup decision across CV
}
\description{
This function finds the majority voted subgroup decision across CV
}
