% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iCF_MAJORITY_VOTE.R
\name{CVBIAS_D_MAJORITY}
\alias{CVBIAS_D_MAJORITY}
\title{This function returns the mean MSE value of those majority voted subgroup decisions across CV}
\usage{
CVBIAS_D_MAJORITY(vote_D_subgroup.L, stability_D_T_r, K, metric)
}
\arguments{
\item{vote_D_subgroup.L}{list of voted subgroup decisions, of those not majorityed, their MSE will be ignored}

\item{metric}{list of metric including RMSE, Rsquared, MAE for continuous variables}
}
\value{
The the mean MSE value across CV
}
\description{
This function returns the mean MSE value of those majority voted subgroup decisions across CV
}
