% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iCF_MAJORITY_VOTE.R
\name{CVBIAS_MAIN}
\alias{CVBIAS_MAIN}
\title{This function returns the mean MSE value across CV}
\usage{
CVBIAS_MAIN(metric)
}
\arguments{
\item{metric}{list of metric including RMSE, Rsquared, MAE for continuous variables}
}
\value{
The the mean MSE value across CV
}
\description{
This function returns the mean MSE value across CV
}
