% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iCF_grplasso.R
\name{DATA4grplasso}
\alias{DATA4grplasso}
\title{Function that prepare training or testing data for model selection based on D4,D3,and D2 subgroup decisions}
\usage{
DATA4grplasso(V_D5_subgroup, V_D4_subgroup, V_D3_subgroup, V_D2_subgroup, dat)
}
\arguments{
\item{V_D4_subgroup}{subgroup decision from depth 4 CF}

\item{V_D3_subgroup}{subgroup decision from depth 3 CF}

\item{V_D2_subgroup}{subgroup decision from depth 2 CF}

\item{dat}{the dataset}
}
\value{
the data and formula for group lasso
}
\description{
Function that prepare training or testing data for model selection based on D4,D3,and D2 subgroup decisions
}
