% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iCF_SUBGROUP_DECISION.R
\name{DECI_treeN1_TREE_TF}
\alias{DECI_treeN1_TREE_TF}
\title{Function that check if the tree of the subgroup decision has the correct top node}
\usage{
DECI_treeN1_TREE_TF(decision, vote_sg, vote_t, tree_true)
}
\arguments{
\item{decision}{the subgroup decision}

\item{vote_sg}{}

\item{tree_true}{true tree structure}
}
\value{
subgroup decision
}
\description{
Function that check if the tree of the subgroup decision has the correct top node
}
