% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iCF_SUBGROUP_ANALYSIS.R
\name{DELTA_Y_DATA}
\alias{DELTA_Y_DATA}
\title{Function that extract ATE, ATT, and crude Delta Y and the coeffecient of GLM for ggploting error bar of Delta Y for both training and testing data.
Run SUBSETTING, ANA_SUBGROUP, and EXTRACT_SUB_ANA 3 functions
for Interaction Tree}
\usage{
DELTA_Y_DATA(Deci, traindata, testdata, methodL)
}
\arguments{
\item{Deci_IT}{the decision from IT}
}
\value{
The list of extracted ATE (IPTW), ATT (SMR), and crude treatment effect (i.e., Delta Y) for both training and testing data, and model df of coeffecient for treatment for ATE, ATT, and crude
}
\description{
Function that extract ATE, ATT, and crude Delta Y and the coeffecient of GLM for ggploting error bar of Delta Y for both training and testing data.
Run SUBSETTING, ANA_SUBGROUP, and EXTRACT_SUB_ANA 3 functions
for Interaction Tree
}
