% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iCF_SUBGROUP_ANALYSIS.R
\name{DELTA_Y_DATA_mCF}
\alias{DELTA_Y_DATA_mCF}
\title{Function that extract ATE, ATT, and crude Delta Y and the coeffecient of GLM for ggploting error bar of Delta Y for both training and testing data.
Run SUBSETTING, ANA_SUBGROUP, and EXTRACT_SUB_ANA 3 functions
Forest could be oneCF, iCF, or iCFv}
\usage{
DELTA_Y_DATA_mCF(Deci_D4, Deci_D3, Deci_D2, traindata, testdata, methodL)
}
\arguments{
\item{Deci_D4}{the voted decision from D4 forest}

\item{Deci_D3}{the voted decision from D3 forest}

\item{Deci_D2}{the voted decision from D2 forest}
}
\value{
The list of extracted ATE (IPTW), ATT (SMR), and crude treatment effect (i.e., Delta Y) for both training and testing data, and model df of coeffecient for treatment for ATE, ATT, and crude
}
\description{
Function that extract ATE, ATT, and crude Delta Y and the coeffecient of GLM for ggploting error bar of Delta Y for both training and testing data.
Run SUBSETTING, ANA_SUBGROUP, and EXTRACT_SUB_ANA 3 functions
Forest could be oneCF, iCF, or iCFv
}
