% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iCF_SUBGROUP_DECISION.R
\name{DETECT_STRING}
\alias{DETECT_STRING}
\title{Function check if vector string in 2nd position nested in the list in 1st position}
\usage{
DETECT_STRING(your_list, vector_strings)
}
\arguments{
\item{your_list}{list}

\item{vector_strings}{vector of string}
}
\value{
T of F
}
\description{
Function check if vector string in 2nd position nested in the list in 1st position
}
