% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iCF_SUBGROUP_DECISION.R
\name{DISCOVERYRATE}
\alias{DISCOVERYRATE}
\title{Function that assess accuracy, false positive, and false negative.
Accuracy definition: all subgup decisions (for tree-based methods that explicitly give subgroup) or all interactions (for non-tree based methds) are accurately identified}
\usage{
DISCOVERYRATE(
  decision,
  truth_description,
  tree_true_subgroup,
  truth_INT,
  method,
  metric
)
}
\arguments{
\item{decision}{the final decision could be subgroup dataframe or interction vector}

\item{truth_description}{the description of truth}

\item{tree_true_subgroup}{the true subgroup}

\item{truth_INT}{the true interaction, when the metric==CATEmax, use true CATEmax subgroup instead of true interaction!!!}

\item{method}{different methods (tree-based or non-tree-based)}

\item{metric}{assesing perforamnce of methods by subgroup or interaction identification}
}
\value{
value=score of accuracy assessment; sg_Nacc_Nall=\% of true subgroups among identified subgroups; sg_Nacc_NallT=\% of true subgroups among total true subgroups
}
\description{
Function that assess accuracy, false positive, and false negative.
Accuracy definition: all subgup decisions (for tree-based methods that explicitly give subgroup) or all interactions (for non-tree based methds) are accurately identified
}
