% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iCF_SUBGROUP_ANALYSIS.R
\name{DltY_DATA}
\alias{DltY_DATA}
\title{Function that obtain residual and MSE for each observation}
\usage{
DltY_DATA(decision, dataset, label)
}
\arguments{
\item{decision}{the subgroup decision by different methods}

\item{dataset}{the dataset analyzed for residual & MSE}

\item{label}{if residual and MSE is for true effect or observed effect}
}
\value{
the dataframe of residual & MSE
}
\description{
Function that obtain residual and MSE for each observation
}
