% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iCF_SUBGROUP_ANALYSIS.R
\name{DLTYCOMBO}
\alias{DLTYCOMBO}
\title{Function that combine Delt_Y and its confidence interval, 2 sample test (W=1 vs W=0) in each subgroups}
\usage{
DLTYCOMBO(DltY_iCF)
}
\arguments{
\item{DltY_iCF}{list containing ATT, ATE of D4,3,2 subgroups in training and testing data}
}
\value{
the dataset that combines Delt_Y and its confidence interval, 2 sample test (W=1 vs W=0) in each subgroups
}
\description{
Function that combine Delt_Y and its confidence interval, 2 sample test (W=1 vs W=0) in each subgroups
}
