% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_tree_MSegar.R
\name{estimate_params}
\alias{estimate_params}
\title{Function that calculates the means and variances of leaves of a pruned tree, using out-of-bag samples.}
\usage{
estimate_params(X, Y, tree, prune_info)
}
\arguments{
\item{X}{Covariates}

\item{Y}{Prediction values of the tree. Could just be Y for regression forest, could be fitted tau}

\item{tree}{The tree that we want to calculate with.}

\item{prune_info}{Info about where the tree is pruned.}
}
\value{
A list, tree_with_oob, where each element represents a node in the tree, with oob samples, oob means, oob standard deviations
}
\description{
Function that calculates the means and variances of leaves of a pruned tree, using out-of-bag samples.
}
\examples{
\dontrun{
# Train a regression forest
n <- 50; p <- 10
X <- matrix(rnorm(n*p), n, p)
Y <- X[,1] * rnorm(n)
forest <- grf::regression_forest(X,Y)

# Find the best tree
best_tree_info <- find_best_tree(Y, forest)

# Calculate parameters for the best pruned tree
best_tree_params <- estimate_params(X, Y, get_tree(forest,best_tree_info$best_tree), best_tree_info$best_prune_info)
}

}
