% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iCF_SUBGROUP_ANALYSIS.R
\name{EXTRACT_SUB_ANA}
\alias{EXTRACT_SUB_ANA}
\title{Function that extract calcualts observed treatment effect size from each depth of iCF and saved in a separate dataframe
In order to record data from each simulation}
\usage{
EXTRACT_SUB_ANA(Subgroup_dltY_D, Depth, datatype, method)
}
\arguments{
\item{Subgroup_dltY_D}{the list of stratified population accroding to subgroup decision}
}
\value{
The dataframe of ATE (IPTW), ATT (SMR), and crude treatment effect (i.e., Delta Y)
}
\description{
Function that extract calcualts observed treatment effect size from each depth of iCF and saved in a separate dataframe
In order to record data from each simulation
}
