% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iCF_SUBGROUP_DECISION.R
\name{FINALDECI}
\alias{FINALDECI}
\title{Function that does shows the final subgroup decisions}
\usage{
FINALDECI(grp_pick, V_D4_subgroup, V_D3_subgroup, V_D2_subgroup)
}
\arguments{
\item{grp_pick}{P-value for raw full CF}

\item{iterationNo}{}
}
\value{
the final subgroup decision,
}
\description{
Function that does shows the final subgroup decisions
}
