% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_tree_MSegar.R
\name{find_leaf}
\alias{find_leaf}
\title{Takes the x (covariates) of a data point, and finds the leaf in the pruned tree that it belongs to}
\usage{
find_leaf(x, tree, prune_info)
}
\arguments{
\item{x}{Covariates of a data point}

\item{tree}{The tree that we want to put data points into}

\item{prune_info}{Info about which are leaves in the pruned tree}
}
\value{
The leaf that the data point belongs to.
}
\description{
Takes the x (covariates) of a data point, and finds the leaf in the pruned tree that it belongs to
}
