% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iCF_SUBGROUP_DECISION.R
\name{FL_Model_Selection}
\alias{FL_Model_Selection}
\title{Function that does backward selection of best model, works for both continuous (partial F-test) and binary outcome (likelihood ratio test)}
\usage{
FL_Model_Selection(P_basicvsg2, P_g2vsg23, P_g23vsg234, siglevel)
}
\arguments{
\item{P_basicvsg2}{P-value for partial F-test or LRT for main effect model vs D2 interaction term model}

\item{P_g2vsg23}{P-value for partial F-test or LRT for D2 interaction term model vs D3 and D2 interaction terms model}

\item{P_g23vsg234}{P-value for partial F-test or LRT for D3 and D2 interaction term model vs D4, D3 and D2 interaction terms model}

\item{siglevel}{predefined significance level for P-value}
}
\value{
the data for group lasso
}
\description{
Function that does backward selection of best model, works for both continuous (partial F-test) and binary outcome (likelihood ratio test)
}
