% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_GenSimData.R
\name{GEN_SIM_DATA_HTE}
\alias{GEN_SIM_DATA_HTE}
\title{This function generate simulated data, may adjust parameter to check different effect size}
\usage{
GEN_SIM_DATA_HTE(
  scenario,
  intTRUE,
  outcome_type,
  B1,
  B2,
  B3,
  B8,
  B9,
  B1_2,
  B1_3,
  B1_8,
  B2_3,
  B3_8,
  B1_9,
  B8_9,
  B3_9,
  B1_2_3,
  B1_3_8,
  B1_3_9,
  B1_8_9,
  B3_8_9,
  B1_3_8_9,
  A1_2,
  A1_3,
  A1_8,
  A2_3,
  A3_8,
  A1_9,
  A8_9,
  A3_9,
  nstudy,
  seed = NULL
)
}
\arguments{
\item{scenario}{scenario for propensity score model}

\item{intTRUE}{scenario for HTE}

\item{nstudy}{sample size}

\item{a8}{coefficient for W:X1}

\item{a9}{coefficient for W:X3}

\item{a10}{coefficient for X1:X3}

\item{a11}{coefficient for W:X1:X3}

\item{a12}{coefficient for W:X8}

\item{a14}{coefficient for X3:X8}

\item{a15}{coefficient for W:X1:X8}

\item{a16}{coefficient for W:X3:X8}

\item{a17}{coefficient for X1:X3:X8}

\item{a18}{coefficient for W:X1:X3:X8}

\item{a_5_1_3_9}{coefficient for W:X1:X3:X9, non-zero when adding 5-way interaction, otherwise 0}

\item{a_5_1_8_9}{coefficient for W:X1:X8:X9, non-zero when adding 5-way interaction, otherwise 0}

\item{a_5_3_8_9}{coefficient for W:X3:X8:X9, non-zero when adding 5-way interaction, otherwise 0}

\item{a21}{coefficient for W:X2}

\item{a22}{coefficient for X2:X3}

\item{a23}{coefficient for W:X2:X3}

\item{a2s2}{coefficient for I(X2^2)}

\item{a4s2}{coefficient for I(X4^2)}

\item{a1i3}{coefficient for non-additivity X1:X3, if a1i3!=0 then a10=0}

\item{a2i4}{coefficient for non-additivity X2:X4}

\item{a2i3}{coefficient for non-additivity X2:X3, if a2i3!=0 then a22=0}

\item{a3i4}{coefficient for non-additivity X3:X4}

\item{a_5_9}{coefficient for W:X9, non-zero when adding 5-way interaction, otherwise 0}

\item{a_5_1_9}{coefficient for W:X1:X9, non-zero when adding 5-way interaction, otherwise 0}

\item{a_5_8_9}{coefficient for W:X8:X9, non-zero when adding 5-way interaction, otherwise 0}

\item{a_5_3_9}{coefficient for W:X3:X9, non-zero when adding 5-way interaction, otherwise 0}

\item{a_5_1_3_8_9}{coefficient for W:X1:X3:X8:X9, non-zero when adding 5-way interaction, otherwise 0}

\item{a_5_19}{coefficient for X1:X9, non-zero when adding 5-way interaction, otherwise 0}

\item{a_5_89}{coefficient for X8:X9, non-zero when adding 5-way interaction, otherwise 0}

\item{a_5_39}{coefficient for X3:X9, non-zero when adding 5-way interaction, otherwise 0}

\item{a_5_139}{coefficient for X1:X3:X9, non-zero when adding 5-way interaction, otherwise 0}

\item{a_5_189}{coefficient for X1:X8:X9, non-zero when adding 5-way interaction, otherwise 0}

\item{a_5_389}{coefficient for X3:X8:X9, non-zero when adding 5-way interaction, otherwise 0}
}
\value{
The subsetted list that all are identical to reference list.
}
\description{
This function generate simulated data, may adjust parameter to check different effect size
}
