% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iCF_SUBGROUP_ANALYSIS.R
\name{GET_SUBGROUP_ID}
\alias{GET_SUBGROUP_ID}
\title{Function that obtain subgroup ID}
\usage{
GET_SUBGROUP_ID(decision, dataset)
}
\arguments{
\item{decision}{the subgroup decision by different methods}

\item{dataset}{the dataset analyzed for residual & MSE}
}
\value{
the dataframe of residual & MSE
}
\description{
Function that obtain subgroup ID
}
