% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iCF_TREE_build.R
\name{GET_TREE_L}
\alias{GET_TREE_L}
\title{This function connects the majority vote (or any) deicison with the list of original tree in list format}
\usage{
GET_TREE_L(iCF_D, listindex)
}
\arguments{
\item{Maj_Tree_Decision}{a majority voted tree decision df that need to be presented in a tree plot}

\item{ori_TreeL}{#a list of original tree list}
}
\value{
The original tree (lsit) that match the voted tree decision (df)
}
\description{
This function connects the majority vote (or any) deicison with the list of original tree in list format
}
