% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GG_toolbox.R
\name{GG_SG_DELTA_Y}
\alias{GG_SG_DELTA_Y}
\title{Function that shows subgroup Delta Y by subgrooup decisions obtained from each depth of forest (oneCF, iCF, or iCFv)}
\usage{
GG_SG_DELTA_Y(model_df, tree_depth, legend_para, splitter)
}
\arguments{
\item{model_df}{the df of coefficient (i.e. Delta Y) and its CI}
}
\value{
The ggplot
}
\description{
Function that shows subgroup Delta Y by subgrooup decisions obtained from each depth of forest (oneCF, iCF, or iCFv)
}
