% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iCF_TREE_build.R
\name{GRID_LEAFSIZE}
\alias{GRID_LEAFSIZE}
\title{This function find right leaf size at each depth}
\usage{
GRID_LEAFSIZE(
  depth,
  treeNo,
  iterationNo,
  initial,
  iCF_D_BT_exp_SG_D_mean,
  tune_unit
)
}
\arguments{
\item{depth}{depth of a forest}

\item{initial}{intially tried value (the denominator of N/initial, e.g. minimum-leaf-size=N/25 for D2 tree, initial=25}

\item{upper}{upper level if we need to get a deeper tree}

\item{lower}{lower level if we need to get a shallower tree}
}
\value{
the tuned leaf size
}
\description{
This function find right leaf size at each depth
}
