% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iCF_TREE_build.R
\name{hdiCF_VS_incremental}
\alias{hdiCF_VS_incremental}
\title{This function is the wrapper function to run raw causal forest and for initial variable selection for hdICF,
which will be applied to a list of selected X by incremetally trimming VI score}
\usage{
hdiCF_VS_incremental(X_selected)
}
\arguments{
\item{X_selected}{selected variables index}

\item{Train_cf}{training data}
}
\value{
the accuracy of predicting W from raw causal forest with different selected hdX
}
\description{
This function is the wrapper function to run raw causal forest and for initial variable selection for hdICF,
which will be applied to a list of selected X by incremetally trimming VI score
}
