% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iCF_CV.R
\name{iCFCV}
\alias{iCFCV}
\title{function to run Cross-Validation on whole dataset (except for external validation set) to get final subgroup decision G_iCF}
\usage{
iCFCV(
  dat,
  K,
  treeNo,
  iterationNo,
  min.split.var,
  split_val_round_posi,
  P_threshold,
  variable_type,
  hdpct,
  HTE_P_cf.raw
)
}
\arguments{
\item{dat}{dataset for overall population}

\item{K}{cross-validation fold}
}
\value{
final subgroup decision G_iCF
}
\description{
function to run Cross-Validation on whole dataset (except for external validation set) to get final subgroup decision G_iCF
}
