% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_GenSimData.R
\name{inv.logit}
\alias{inv.logit}
\title{This function returns the inverse logit of a given numeric value, i.e., translate it into probabilities between 0 and 1.}
\usage{
inv.logit(m)
}
\arguments{
\item{m}{the model of event Y (binary outcome), m = A + B*X}
}
\value{
The probability of Y
}
\description{
This function returns the inverse logit of a given numeric value, i.e., translate it into probabilities between 0 and 1.
}
