% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iCF_TREE_build.R
\name{LEAFSIZE_tune}
\alias{LEAFSIZE_tune}
\title{This function tune leaf size for iCF}
\usage{
LEAFSIZE_tune(Ntrain, initial_D2, treeNo, iterationNo, tune_unit)
}
\arguments{
\item{Ntrain}{sample size}

\item{treeNo}{tree No, need to be >= 50, if =10 has this error: Error in lm.wfit(x, y, w, offset = offset, singular.ok = singular.ok,  : 0 (non-NA) cases}

\item{iterationNo}{iteration No}
}
\value{
the tuned leaf size
}
\description{
This function tune leaf size for iCF
}
