% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iCF_MAJORITY_VOTE.R
\name{MAJORITY_COUNT}
\alias{MAJORITY_COUNT}
\title{This function count the occurence of majority}
\usage{
MAJORITY_COUNT(list_ref, list_all)
}
\arguments{
\item{list_ref}{list of unique trees, tree relaxed, or subgroups}

\item{list_all}{prepared list of trees, tree relaxed, or subgroups}
}
\value{
The occurency of majority
}
\description{
This function count the occurence of majority
}
