% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iCF_MAJORITY_VOTE.R
\name{MAJORITY_PICK_LEAF}
\alias{MAJORITY_PICK_LEAF}
\title{This function pick those match matjority vote and with LEAF or NODE}
\usage{
MAJORITY_PICK_LEAF(inputlist, leaf_para)
}
\arguments{
\item{inputlist}{a list of df match the voted}

\item{leaf_para}{"TRUE", means it is a leaf}
}
\value{
The subsetted list that match voted
}
\description{
This function pick those match matjority vote and with LEAF or NODE
}
