% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iCF_MAJORITY_VOTE.R
\name{MAJORITY_VOTE}
\alias{MAJORITY_VOTE}
\title{This function performs majority vote among best trees from iterative CF}
\usage{
MAJORITY_VOTE(
  list0,
  list1,
  list2,
  list3,
  truth,
  truth_N1,
  truth_N123,
  split_val_round_posi
)
}
\arguments{
\item{list0}{original list of trees}

\item{list1}{prepared list of trees, tree relaxed, or subgroups}

\item{list2}{only useful for prepared list if by subgroup, then remove node column to count for majority.tree}

\item{list3}{original list of split frequency!}

\item{truth}{truth}

\item{truth_N1}{true top node for tree}

\item{truth_N123}{true tope 3 nodes for tree}

\item{split_val_round_posi}{rounding parameter for decimal place}
}
\value{
The subsetted list that all are identical to reference list.
}
\description{
This function performs majority vote among best trees from iterative CF
}
