% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iCF_SUBGROUP_ANALYSIS.R
\name{MAKE_GROUP_COLUMN}
\alias{MAKE_GROUP_COLUMN}
\title{Function that make intearction term of W:subgroups(dummy varialbe)}
\usage{
MAKE_GROUP_COLUMN(dat, depth, dummytype, level)
}
\arguments{
\item{dat}{the dataset with subgroup ID and defintion for each observation}

\item{depth}{the label describe type (depth) of iCF dataset}

\item{dummytype}{the column created could be W:group, group, or group&W:group}

\item{level}{if level="g", then make dummy variable for each specific subgroup g; if level="G" then rename "SubgroupID" as "G_i"}
}
\value{
the dataset with interaction terms of W:subgroups
}
\description{
Function that make intearction term of W:subgroups(dummy varialbe)
}
