% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iCF_TREE_build.R
\name{MinLeafSizeTune}
\alias{MinLeafSizeTune}
\title{This function for manually tuning MLS by inputting denominator (divide sample size into) and show plot of depth}
\usage{
MinLeafSizeTune(
  dat,
  denominator,
  treeNo,
  iterationNo,
  split_val_round_posi,
  depth,
  color
)
}
\arguments{
\item{depth}{depth descripion, e.g. "D4"}

\item{demoninator}{a demoninator to devide the sample}
}
\value{
The original tree (lsit) that match the voted tree decision (df)
}
\description{
This function for manually tuning MLS by inputting denominator (divide sample size into) and show plot of depth
}
