% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iCF_SUBGROUP_ANALYSIS.R
\name{MSE_DltY}
\alias{MSE_DltY}
\title{Function that extract MSE from MSE dataframe statify population and saved stratified population (subgroup) in a list.}
\usage{
MSE_DltY(DltY_true, DltY)
}
\arguments{
\item{DltY}{the dataframe of MSE obtained from function DltY_DATA}
}
\value{
the list of MSE, residual by IPTW (ate) or SMR (att) weighting
}
\description{
Function that extract MSE from MSE dataframe statify population and saved stratified population (subgroup) in a list.
}
