% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iCF_SUBGROUP_DECISION.R
\name{OTHERTREE_DECI}
\alias{OTHERTREE_DECI}
\title{Function that obtain subgroup decision from other tree-based method such as IT tree}
\usage{
OTHERTREE_DECI(vote_subgroup)
}
\arguments{
\item{vote_subgroup}{the subgroup decision from voted tree structure}
}
\value{
subgroup decision
}
\description{
Function that obtain subgroup decision from other tree-based method such as IT tree
}
