% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iCF_SUBGROUP_DECISION.R
\name{PICK_CV}
\alias{PICK_CV}
\title{Function that does shows the depth of subgroup selected}
\usage{
PICK_CV(
  HTE_P_value,
  MAF_m,
  MAF_g2,
  MAF_g3,
  MAF_g4,
  MAF_g5,
  V_D2_SG.L,
  V_D3_SG.L,
  V_D4_SG.L,
  V_D5_SG.L,
  S_D2_T_r,
  S_D3_T_r,
  S_D4_T_r,
  S_D5_T_r,
  P_threshold,
  K
)
}
\arguments{
\item{HTE_P_value}{P-value for raw full CF}

\item{MAF_m}{the mean MSE value across CV for main effect model}

\item{MAF_g2}{the mean MSE value across CV for D2 SG model}

\item{MAF_g3}{the mean MSE value across CV for D3 SG model}

\item{MAF_g4}{the mean MSE value across CV for D4 SG model}

\item{MAF_g5}{the mean MSE value across CV for D5 SG model}

\item{V_D2_SG.L}{the mean MSE value across CV for D2 SG model}

\item{V_D3_SG.L}{the mean MSE value across CV for D3 SG model}

\item{V_D4_SG.L}{the mean MSE value across CV for D4 SG model}

\item{V_D5_SG.L}{the mean MSE value across CV for D5 SG model}

\item{S_D2_T_r}{the mean MSE value across CV for D2 SG model}

\item{S_D3_T_r}{the mean MSE value across CV for D3 SG model}

\item{S_D4_T_r}{the mean MSE value across CV for D4 SG model}

\item{S_D5_T_r}{the mean MSE value across CV for D5 SG model}

\item{P_threshold}{P-value threshold}

\item{K}{cross-validation fold}

\item{S_D2_CV}{the mean MSE value across CV for D2 SG model}

\item{S_D3_CV}{the mean MSE value across CV for D3 SG model}

\item{S_D4_CV}{the mean MSE value across CV for D4 SG model}

\item{S_D5_CV}{the mean MSE value across CV for D5 SG model}

\item{S_D2_vote}{the mean MSE value across CV for D2 SG model}

\item{S_D3_vote}{the mean MSE value across CV for D3 SG model}

\item{S_D4_vote}{the mean MSE value across CV for D4 SG model}

\item{S_D5_vote}{the mean MSE value across CV for D5 SG model}
}
\value{
selected SG and it's stability of CV, and stability of Vote
}
\description{
Function that does shows the depth of subgroup selected
}
