% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iCF_SUBGROUP_DECISION.R
\name{PICK_m2345}
\alias{PICK_m2345}
\title{Function that does shows the depth of subgroup selected}
\usage{
PICK_m2345(HTE_P_value, m, g2, g3, g4, g5, P_threshold)
}
\arguments{
\item{HTE_P_value}{P-value for raw full CF}

\item{m}{metric for main effects model}

\item{g2}{metric for g2 model}

\item{g3}{metric for g3 model}

\item{g4}{metric for g4 model}

\item{g5}{metric for g5 model}
}
\value{
the depth of subgroup selected, e.g. "W:G2x"
}
\description{
Function that does shows the depth of subgroup selected
}
