% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iCF_SUBGROUP_ANALYSIS.R
\name{prop.func}
\alias{prop.func}
\title{Function that predict PS using LASSO algorithm}
\usage{
prop.func(x, trt)
}
\arguments{
\item{x}{covariates}

\item{trt}{"treatment"}
}
\value{
the prediced propensity score
}
\description{
Function that predict PS using LASSO algorithm
}
