% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_tree_MSegar.R
\name{r_loss}
\alias{r_loss}
\title{Function that takes in a list of samples, with the corresponding Y values, and calculates the r_loss, assuming a regression tree.
If a leaf only has one sample, then we return the Y value squared. We choose this value because
we don't want to return 0, which will encourage the tree to keep leaves with 1 sample, but we also
don't want to return a huge value, which will force the tree to never have leaves with 1 sample, even if
the sample is an outlier. But of course we could choose to return something else when there is only
one sample in the leaf, if it makes more sense.}
\usage{
r_loss(Y, samples, W.orig = NULL, W.hat = NULL, M.hat = NULL, Tau.hat = NULL)
}
\arguments{
\item{Y}{The Y values}

\item{samples}{The samples on which to calculate the r_loss.}

\item{W.orig}{W values that indicate treatment; used for causal forest.}

\item{W.hat}{Fitted W.hat for causal forest}

\item{M.hat}{M.hat for causal forest}

\item{Tau.hat}{Fitten tau values for causal forest}
}
\value{
The r_loss of the samples.
}
\description{
Function that takes in a list of samples, with the corresponding Y values, and calculates the r_loss, assuming a regression tree.
If a leaf only has one sample, then we return the Y value squared. We choose this value because
we don't want to return 0, which will encourage the tree to keep leaves with 1 sample, but we also
don't want to return a huge value, which will force the tree to never have leaves with 1 sample, even if
the sample is an outlier. But of course we could choose to return something else when there is only
one sample in the leaf, if it makes more sense.
}
