% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iCF_SUBGROUP_DECISION.R
\name{SG2INT}
\alias{SG2INT}
\title{Function that convert overall subgroup decision into interaction expressions (e.g. subgroup decision as 1: X1<=0; 2: X1>0 will be converted into W:X1 interaction expression)}
\usage{
SG2INT(deci)
}
\arguments{
\item{condition}{a specific subgroup definition (one group), e.g.  X1<=0 & X3<=0}
}
\value{
interaction, e.g. W:X1:W3
}
\description{
Function that convert overall subgroup decision into interaction expressions (e.g. subgroup decision as 1: X1<=0; 2: X1>0 will be converted into W:X1 interaction expression)
}
