% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iCF_SUBGROUP_DECISION.R
\name{SG_CONDITION_2_INT}
\alias{SG_CONDITION_2_INT}
\title{Function that convert each subgroup definition into interaction (e.g. a subgroup defined as X1<=0 & X3<=0 will be converted into W:X1:W3)}
\usage{
SG_CONDITION_2_INT(condition)
}
\arguments{
\item{condition}{a specific subgroup definition (one group), e.g.  X1<=0 & X3<=0}
}
\value{
interaction, e.g. W:X1:W3
}
\description{
Function that convert each subgroup definition into interaction (e.g. a subgroup defined as X1<=0 & X3<=0 will be converted into W:X1:W3)
}
