% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iCF_MAJORITY_VOTE.R
\name{SG_PROBABILITY}
\alias{SG_PROBABILITY}
\title{This function check the probability of the majority voted subgroup from iCF}
\usage{
SG_PROBABILITY(list1)
}
\arguments{
\item{list1}{prepared list of trees, tree relaxed, or subgroups}
}
\value{
The subsetted list that all are identical to reference list.
}
\description{
This function check the probability of the majority voted subgroup from iCF
}
