% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iCF_SUBGROUP_DECISION.R
\name{SGMODEL_DATA}
\alias{SGMODEL_DATA}
\title{Function that does prepare data that
\enumerate{
\item ready to apply the formulas including interacting terms between treatment and factor subgroup deicision variable;
2)transformed outcome
}}
\usage{
SGMODEL_DATA(dat, outcome_type_at)
}
\arguments{
\item{dat}{P-value for raw full CF}

\item{outcome_type}{"actual" or "transform"}
}
\value{
the prepared data and "contr"
}
\description{
Function that does prepare data that
\enumerate{
\item ready to apply the formulas including interacting terms between treatment and factor subgroup deicision variable;
2)transformed outcome
}
}
