% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iCF_TREE_build.R
\name{SPLIT_VAL_IMPUTE_D2}
\alias{SPLIT_VAL_IMPUTE_D2}
\title{This function impute Impute split values for splitters of D2 voted tree from D4 synthetic tree if top 1 node, i.e.,root node, are identical}
\usage{
SPLIT_VAL_IMPUTE_D2(
  level_D2_splitter_1,
  vote_D2_tree.syn.ori,
  vote_D4_tree.syn
)
}
\arguments{
\item{level_D2_splitter_1}{the node 1 (splitter1) of D3 voted tree structure}

\item{vote_D2_tree.syn.ori}{the original vote D3 tree stucutre (without imputing split value)}

\item{vote_D4_tree.syn}{the original vote D4 tree stucutre (without imputing split value)}
}
\value{
The imputed D3 tree if Node 1,2,3 of D3 identical to Node 1,2, 3 of D4, or original D3 tree if Node 1, 2, 3 of D3 not identical to D4
}
\description{
This function impute Impute split values for splitters of D2 voted tree from D4 synthetic tree if top 1 node, i.e.,root node, are identical
}
