% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iCF_TREE_build.R
\name{SPLIT_VAL_IMPUTE_D3}
\alias{SPLIT_VAL_IMPUTE_D3}
\title{This function impute Impute split values for splitters of D3 voted tree from D4 synthetic tree if top 3 nodes are identical}
\usage{
SPLIT_VAL_IMPUTE_D3(
  level_D3_splitter_1,
  level_D3_splitter_2,
  level_D3_splitter_3,
  vote_D3_tree.syn.ori,
  vote_D4_tree.syn
)
}
\arguments{
\item{level_D3_splitter_1}{the node 1 (splitter1) of D3 voted tree structure}

\item{level_D3_splitter_2}{the node 2 (splitter2) of D3 voted tree structure}

\item{level_D3_splitter_3}{the node 3 (splitter3) of D3 voted tree structure}

\item{vote_D3_tree.syn.ori}{the original vote D3 tree stucutre (without imputing split value)}

\item{vote_D4_tree.syn}{the original vote D4 tree stucutre (without imputing split value)}
}
\value{
The imputed D3 tree if Node 1,2,3 of D3 identical to Node 1,2, 3 of D4, or original D3 tree if Node 1, 2, 3 of D3 not identical to D4
}
\description{
This function impute Impute split values for splitters of D3 voted tree from D4 synthetic tree if top 3 nodes are identical
}
