% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iCF_TREE_build.R
\name{SPLITTER_LEVEL}
\alias{SPLITTER_LEVEL}
\title{This function find out a splitter is continuous, ordinal, orbinary variable}
\usage{
SPLITTER_LEVEL(tree, row_index)
}
\arguments{
\item{tree}{the dataframe for a causal tree}

\item{row_index}{#the row number of the dataframe}
}
\value{
The level of this splitter, e.g.if continous splitter, then > 8
}
\description{
This function find out a splitter is continuous, ordinal, orbinary variable
}
