% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iCF_SUBGROUP_DECISION.R
\name{SPLITVAR_SIGN_SPLITVAL_LofDf_VT}
\alias{SPLITVAR_SIGN_SPLITVAL_LofDf_VT}
\title{Function that make subgroup defintion as a list of df, each df representing each condition of this subgroup definition and has 3 key variabs: parent_split_var_, parent_sign_, parent_split_val}
\usage{
SPLITVAR_SIGN_SPLITVAL_LofDf_VT(Deci_VT_con0)
}
\arguments{
\item{Deci_VT_con0}{subgroup definition string with "&" sign come from VirtualTWin}
}
\value{
subgroup defintion as a list of df, each df representing each condition of this subgroup definition
}
\description{
Function that make subgroup defintion as a list of df, each df representing each condition of this subgroup definition and has 3 key variabs: parent_split_var_, parent_sign_, parent_split_val
}
