% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iCF_MAJORITY_VOTE.R
\name{STANDARD_CHECK}
\alias{STANDARD_CHECK}
\title{This function connects if identical or not label to a list and subsetting the list to exclude those not identical to reference list
if_ID means if Identical to MAJORITY VOTED or TRUTH}
\usage{
STANDARD_CHECK(standardLIST, inputLIST)
}
\arguments{
\item{standardLIST}{a label list of being identical or not}

\item{inputLIST}{the list need to be labeled and subsetted}
}
\value{
The subsetted list that all are identical to reference list.
}
\description{
This function connects if identical or not label to a list and subsetting the list to exclude those not identical to reference list
if_ID means if Identical to MAJORITY VOTED or TRUTH
}
