% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iCF_SG_PIPELINE.R
\name{SUBGROUP_PIPELINE}
\alias{SUBGROUP_PIPELINE}
\title{function to get subroup decision G_D, get ready for predicting Y (transformed outcome) in testing set,  and models build from G_D to predict Y*}
\usage{
SUBGROUP_PIPELINE(
  X,
  Y,
  W,
  Y.hat,
  W.hat,
  selected_cf.idx,
  leafsize,
  treeNo,
  iterationNo,
  Ntrain,
  Train_ID,
  Test_ID,
  variable_type,
  HTE_P_cf.raw,
  P_threshold,
  split_val_round_posi
)
}
\arguments{
\item{leafsize}{minimum-node-size tuned}
}
\value{
G_D
}
\description{
function to get subroup decision G_D, get ready for predicting Y (transformed outcome) in testing set,  and models build from G_D to predict Y*
}
