% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iCF_SUBGROUP_ANALYSIS.R
\name{SUBSETTING}
\alias{SUBSETTING}
\title{Function that statify population and saved stratified population (subgroup) in a list.
If decision is from D5, D4, D3, or D2 tree, then it will always split.
If decision is from truth, then it could be "NA", i.e. no HTE. In such scenario, set condition_F  = "W<100", i.e. ONE "group" for all observations.}
\usage{
SUBSETTING(decision, dataset)
}
\arguments{
\item{decision}{the final decision obtained from the decision path}

\item{dataset}{Train or Test dataset}
}
\value{
the list of stratified population accroding to subgroup decision
}
\description{
Function that statify population and saved stratified population (subgroup) in a list.
If decision is from D5, D4, D3, or D2 tree, then it will always split.
If decision is from truth, then it could be "NA", i.e. no HTE. In such scenario, set condition_F  = "W<100", i.e. ONE "group" for all observations.
}
