% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iCF_SUBGROUP_DECISION.R
\name{TREE2SUBGROUP}
\alias{TREE2SUBGROUP}
\title{Wrapper function that convert tree structure to subgroup}
\usage{
TREE2SUBGROUP(tree)
}
\arguments{
\item{tree}{tree structure df}
}
\value{
subgroup in list format
}
\description{
Wrapper function that convert tree structure to subgroup
}
