% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_Truth_tree.R
\name{TREE_PERFORMANCE}
\alias{TREE_PERFORMANCE}
\title{function to assess performance of different tree-based subgrouping methods}
\usage{
TREE_PERFORMANCE(Deci_SG, iterationNo)
}
\arguments{
\item{Deci_SG}{subgroup decision}

\item{iterationNo}{distinguish between oneCF and iCF}
}
\value{
final subgroup decision G_iCF
}
\description{
function to assess performance of different tree-based subgrouping methods
}
