% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iCF_SUBGROUP_DECISION.R
\name{VT2CF_format}
\alias{VT2CF_format}
\title{Function that transform the splitting format of VirtualTwin style to CausalForest style:
\enumerate{
\item binary/ordinal splitter >= n.5 to >n, < n.5 to <= n; 2) continuous splitter: >= n to > n-0.001, < n to <= n-0.001
}}
\usage{
VT2CF_format(Deci_VT_con0)
}
\arguments{
\item{Deci_VT_con0}{subgroup definition string with "&" sign (may come from VirtualTWin or CF)}
}
\value{
subgroup decision
}
\description{
Function that transform the splitting format of VirtualTwin style to CausalForest style:
\enumerate{
\item binary/ordinal splitter >= n.5 to >n, < n.5 to <= n; 2) continuous splitter: >= n to > n-0.001, < n to <= n-0.001
}
}
